chore: move test helpers with network request
