+++ test/libcommonserver/CMakeLists.txt
@@ -51,7 +51,6 @@ set_target_properties(${testcommonserver_NAME} PROPERTIES
 
 target_link_libraries(${testcommonserver_NAME}
         ${libcommonserver_NAME}
-        ${libsyncengine_NAME}
 )
 
 if(WIN32)
