+++ test/test_utility/testhelpers.cpp
@@ -22,11 +22,6 @@
 
 #include "libcommon/utility/utility.h"
 #include "libcommonserver/io/iohelper.h"
-#include "libsyncengine/jobs/network/kDrive_API/createdirjob.h"
-#include "libsyncengine/jobs/network/kDrive_API/deletejob.h"
-#include "libsyncengine/jobs/network/kDrive_API/duplicatejob.h"
-#include "libsyncengine/jobs/network/kDrive_API/movejob.h"
-#include "libsyncengine/jobs/network/kDrive_API/upload/uploadjob.h"
 
 #include <fstream>
 
@@ -176,63 +171,6 @@ void setupLogging() {
     if (!Log::instance(Path2WStr(logFilePath))) {
         assert(false);
     }
-}
-
-NodeId createRemoteDir(const int driveDbId, const NodeId &remoteParentId, const SyncName &name) {
-    CreateDirJob job(nullptr, driveDbId, remoteParentId, name);
-    (void) job.runSynchronously();
-    return job.nodeId();
-}
-
-void editRemoteFile(const int driveDbId, const NodeId &remoteFileId, SyncTime *creationTime /*= nullptr*/,
-                    SyncTime *modificationTime /*= nullptr*/, int64_t *size /*= nullptr*/) {
-    const LocalTemporaryDirectory temporaryDir;
-    const auto tmpFilePath = temporaryDir.path() / ("tmpFile_" + CommonUtility::generateRandomStringAlphaNum(10));
-    testhelpers::generateOrEditTestFile(tmpFilePath);
-
-    const auto timestamp = duration_cast<std::chrono::seconds>(
-            time_point_cast<std::chrono::seconds>(std::chrono::system_clock::now()).time_since_epoch());
-    UploadJob job(nullptr, driveDbId, tmpFilePath, remoteFileId, timestamp.count());
-    (void) job.runSynchronously();
-    if (creationTime) {
-        *creationTime = job.creationTime();
-    }
-    if (modificationTime) {
-        *modificationTime = job.modificationTime();
-    }
-    if (size) {
-        *size = job.size();
-    }
-}
-
-void moveRemoteItem(const int driveDbId, const NodeId &remoteFileId, const NodeId &destinationRemoteParentId,
-                    const SyncName &name /*= {}*/) {
-    MoveJob job(nullptr, driveDbId, {}, remoteFileId, destinationRemoteParentId, name);
-    job.setBypassCheck(true);
-    (void) job.runSynchronously();
-}
-
-NodeId duplicateRemoteItem(const int driveDbId, const NodeId &id, const SyncName &newName) {
-    DuplicateJob job(nullptr, driveDbId, id, newName);
-    (void) job.runSynchronously();
-    return job.nodeId();
-}
-
-void deleteRemoteItem(const int driveDbId, const NodeId &id) {
-    DeleteJob job(driveDbId, id);
-    job.setBypassCheck(true);
-    (void) job.runSynchronously();
-}
-
-SyncPath findLocalFileByNamePrefix(const SyncPath &parentAbsolutePath, const SyncName &namePrefix) {
-    IoError ioError(IoError::Unknown);
-    IoHelper::DirectoryIterator dirIt(parentAbsolutePath, false, ioError);
-    bool endOfDir = false;
-    DirectoryEntry entry;
-    while (dirIt.next(entry, endOfDir, ioError) && !endOfDir) {
-        if (CommonUtility::startsWith(entry.path().filename(), namePrefix)) return entry.path();
-    }
-    return {};
 }
 
 } // namespace KDC::testhelpers
